+++ examples/advanced_training/plot_relative_positioning.py
@@ -314,14 +314,11 @@ cp = Checkpoint(dirname='', f_criterion=None, f_optimizer=None, f_history=None)
 early_stopping = EarlyStopping(patience=10)
 train_acc = EpochScoring(
     scoring='accuracy', on_train=True, name='train_acc', lower_is_better=False)
-valid_acc = EpochScoring(
-    scoring='accuracy', on_train=False, name='valid_acc',
-    lower_is_better=False)
+
 callbacks = [
     ('cp', cp),
     ('patience', early_stopping),
     ('train_acc', train_acc),
-    ('valid_acc', valid_acc)
 ]
 
 clf = EEGClassifier(
@@ -371,7 +368,6 @@ ys1 = ['train_loss', 'valid_loss']
 ys2 = ['train_acc', 'valid_acc']
 styles = ['-', ':']
 markers = ['.', '.']
-
 
 fig, ax1 = plt.subplots(figsize=(8, 3))
 ax2 = ax1.twinx()