+++ docs/whats_new.rst
@@ -55,6 +55,7 @@ Enhancements
 - Allow passing only the name of a braindecode model to :class:`braindecode.EEGClassifier` and :class:`braindecode.EEGRegressor` (:gh:`528` by `Pierre Guetschel`_)
 - Standardizing models' last layer names (:gh:`520` by `Bruna Lopes`_ and `Pierre Guetschel`_)
 - Add basic training example with MNE epochs (:gh:`539` by `Pierre Guetschel`_)
+- Log validation accuracy in :class:`braindecode.EEGClassifier` (:gh:`541` by `Pierre Guetschel`_)
 
 Bugs
 ~~~~